+++ pg_metadata/install/sql/upgrade/upgrade_to_1.2.3.sql
@@ -16,13 +16,29 @@ INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item
 INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (142, 'link.mime', 'directory', 'inode/directory', 'Directory (not an official MIME type)',  NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL);
 INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (143, 'link.type', 'ESRI:SHP', 'ESRI Shapefile', 'Vector layer in Shapefile format (.shp)', NULL, NULL, NULL, NULL, NULL, NULL, NULL, 'ESRI Shapefile', 'Vektorlayer im Shapefile-Format (.shp)');
 
+
 -- codes for unknown license and confidentiality
 
 INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (144, 'dataset.license', 'NO', 'No or unknown license', 'Dataset has been published explicitly without license or no license conditions have been documented', 100, NULL, NULL, NULL, NULL, NULL, NULL, 'Keine oder unbekannte Lizenz', 'Die Daten wurden explizit ohne Lizenz freigegeben oder die Lizenz ist nicht bekannt') ON CONFLICT DO NOTHING;
 INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (145, 'dataset.confidentiality', 'UNK', 'Unknown', 'Access restrictions for this dataset are not known', 30, NULL, NULL, NULL, NULL, NULL, NULL, 'Unbekannt', 'Zugriffsbeschränkungen für diese Daten sind nicht bekannt') ON CONFLICT DO NOTHING;
 
 
-SELECT pg_catalog.setval('pgmetadata.glossary_id_seq', 145, true);
+-- remaining INSPIRE contact roles
+
+INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (146, 'contact.contact_role', 'AU', 'Author', 'Party who authored the resource', 45, 'Auteur', 'Partie qui est l’auteur de la ressource', NULL, NULL, NULL, NULL, 'Autor', 'Verfasser der Ressource');
+INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (147, 'contact.contact_role', 'PC', 'Point of Contact', 'Party who can be contacted for acquiring knowledge about or acquisition of the resource', 5, 'Point de contact', 'Partie qu’il est possible de contacter pour s’informer sur la ressource ou en faire l’acquisition', NULL, NULL, NULL, NULL, 'Ansprechpartner', 'Kontakt für Informationen zur Ressource oder deren Bezugsmöglichkeiten');
+INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (148, 'contact.contact_role', 'PI', 'Principal Investigator', 'Key party responsible for gathering information and conducting research', 47, 'Maître d’œuvre', 'Principale partie chargée de recueillir des informations et de mener les recherches', NULL, NULL, NULL, NULL, 'Projektleitung', 'Person oder Stelle, die verantwortlich für die Erhebung der Daten und die Untersuchung ist');
+INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (149, 'contact.contact_role', 'PU', 'Publisher', 'Party who published the resource', 15, 'Éditeur', 'Partie qui a publié la ressource', NULL, NULL, NULL, NULL, 'Herausgeber', 'Person oder Stelle, welche die Ressource veröffentlicht');
+INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (150, 'contact.contact_role', 'RP', 'Resource Provider', 'Party that supplies the resource.', 25, 'Fournisseur de la ressource', 'Partie qui fournit la ressource', NULL, NULL, NULL, NULL, 'Ressourcenanbieter', 'Anbieter der Ressource');
+INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (151, 'contact.contact_role', 'US', 'User', 'Party who uses the resource', 80, 'Utilisateur', 'Partie qui utilise la ressource', NULL, NULL, NULL, NULL, 'Nutzer', 'Nutzer der Ressource');
+
+
+-- UK Open Government Licence
+
+INSERT INTO pgmetadata.glossary (id, field, code, label_en, description_en, item_order, label_fr, description_fr, label_it, description_it, label_es, description_es, label_de, description_de) VALUES (152, 'dataset.license', 'OGL-UK-3.0', 'Open Government Licence v3.0', 'https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/', 85, NULL, NULL, NULL, NULL, NULL, NULL, 'Open Government Licence v3.0', NULL);
+
+
+SELECT pg_catalog.setval('pgmetadata.glossary_id_seq', 152, true);
 
 
 -- new item_order for existing publication frequencies